ci(github): connect npm release with more checks
